codex: address retry review structure feedback (#25147)
