+++ codex-rs/rmcp-client/src/rmcp_client.rs
@@ -14,7 +14,6 @@ use anyhow::anyhow;
 use codex_api::SharedAuthProvider;
 use codex_client::maybe_build_rustls_client_config_with_custom_ca;
 use codex_config::types::McpServerEnvVar;
-use codex_exec_server::ExecServerError;
 use codex_exec_server::HttpClient;
 use futures::FutureExt;
 use futures::future::BoxFuture;
@@ -50,7 +49,6 @@ use rmcp::transport::auth::AuthClient;
 use rmcp::transport::auth::AuthError;
 use rmcp::transport::auth::OAuthState;
 use rmcp::transport::streamable_http_client::StreamableHttpClientTransportConfig;
-use rmcp::transport::streamable_http_client::StreamableHttpError;
 use serde::Deserialize;
 use serde::Serialize;
 use serde_json::Value;
@@ -62,7 +60,6 @@ use tracing::warn;
 
 use crate::elicitation_client_service::ElicitationClientService;
 use crate::http_client_adapter::StreamableHttpClientAdapter;
-use crate::http_client_adapter::StreamableHttpClientAdapterError;
 use crate::in_process_transport::InProcessTransportFactory;
 use crate::load_oauth_tokens;
 use crate::oauth::OAuthPersistor;
@@ -75,9 +72,9 @@ use crate::utils::apply_default_headers;
 use crate::utils::build_default_headers;
 use codex_config::types::OAuthCredentialsStoreMode;
 
-const MCP_CLIENT_INITIALIZE_METRIC: &str = "codex.mcp.client.initialize";
-const JSON_RPC_INTERNAL_ERROR_CODE: i64 = -32603;
-const STREAMABLE_HTTP_RETRY_DELAYS_MS: [u64; 2] = [250, 1_000];
+mod streamable_http_retry;
+
+use self::streamable_http_retry::HandshakeError;
 
 enum PendingTransport {
     InProcess {
@@ -93,63 +90,6 @@ enum PendingTransport {
         transport: StreamableHttpClientTransport<AuthClient<StreamableHttpClientAdapter>>,
         oauth_persistor: OAuthPersistor,
     },
-}
-
-impl PendingTransport {
-    fn is_streamable_http(&self) -> bool {
-        matches!(
-            self,
-            PendingTransport::StreamableHttp { .. }
-                | PendingTransport::StreamableHttpWithOAuth { .. }
-        )
-    }
-
-    fn metric_transport(&self) -> &'static str {
-        match self {
-            PendingTransport::InProcess { .. } => "in_process",
-            PendingTransport::Stdio { .. } => "stdio",
-            PendingTransport::StreamableHttp { .. }
-            | PendingTransport::StreamableHttpWithOAuth { .. } => "streamable_http",
-        }
-    }
-}
-
-fn initialize_metric_tags(
-    transport: &'static str,
-    outcome: &'static str,
-    attempts: usize,
-    retry_exhausted: bool,
-    failure_kind: &'static str,
-) -> Vec<(&'static str, String)> {
-    let attempts = attempts.max(1);
-    vec![
-        ("transport", transport.to_string()),
-        ("outcome", outcome.to_string()),
-        ("retried", (attempts > 1).to_string()),
-        ("attempts", attempts.to_string()),
-        ("retry_count", attempts.saturating_sub(1).to_string()),
-        ("retry_exhausted", retry_exhausted.to_string()),
-        ("failure_kind", failure_kind.to_string()),
-    ]
-}
-
-fn emit_initialize_metric(
-    transport: &'static str,
-    outcome: &'static str,
-    attempts: usize,
-    retry_exhausted: bool,
-    failure_kind: &'static str,
-) {
-    let Some(metrics) = codex_otel::global() else {
-        return;
-    };
-
-    let tags = initialize_metric_tags(transport, outcome, attempts, retry_exhausted, failure_kind);
-    let tag_refs: Vec<(&str, &str)> = tags
-        .iter()
-        .map(|(key, value)| (*key, value.as_str()))
-        .collect();
-    let _ = metrics.counter(MCP_CLIENT_INITIALIZE_METRIC, /*inc*/ 1, &tag_refs);
 }
 
 enum ClientState {
@@ -275,34 +215,6 @@ where
             }
         }
     }
-}
-
-async fn sleep_with_retry_deadline(delay: Duration, deadline: Option<Instant>) -> bool {
-    if let Some(deadline) = deadline {
-        let remaining = deadline.saturating_duration_since(Instant::now());
-        if remaining.is_zero() {
-            return false;
-        }
-        time::timeout(remaining, time::sleep(delay)).await.is_ok()
-    } else {
-        time::sleep(delay).await;
-        true
-    }
-}
-
-#[derive(Debug, thiserror::Error)]
-enum ClientOperationError {
-    #[error(transparent)]
-    Service(#[from] rmcp::service::ServiceError),
-    #[error("timed out awaiting {label} after {duration:?}")]
-    Timeout { label: String, duration: Duration },
-}
-
-#[derive(Debug, thiserror::Error)]
-#[error("handshaking with MCP server failed: {source}")]
-struct HandshakeError {
-    #[source]
-    source: rmcp::service::ClientInitializeError,
 }
 
 pub type Elicitation = CreateElicitationRequestParams;
@@ -942,315 +854,6 @@ impl RmcpClient {
         };
 
         Ok((Arc::new(service), oauth_persistor))
-    }
-
-    async fn connect_pending_transport_with_initialize_retries(
-        &self,
-        initial_transport: PendingTransport,
-        client_service: ElicitationClientService,
-        timeout: Option<Duration>,
-    ) -> Result<(
-        Arc<RunningService<RoleClient, ElicitationClientService>>,
-        Option<OAuthPersistor>,
-    )> {
-        let should_retry = initial_transport.is_streamable_http();
-        let metric_transport = initial_transport.metric_transport();
-        let retry_deadline = timeout.map(|duration| Instant::now() + duration);
-        let mut pending_transport = Some(initial_transport);
-
-        let retry_schedule = STREAMABLE_HTTP_RETRY_DELAYS_MS
-            .iter()
-            .copied()
-            .map(Some)
-            .chain(std::iter::once(None));
-
-        for (attempt, retry_delay_ms) in retry_schedule.enumerate() {
-            let attempt_count = attempt + 1;
-            let attempt_timeout =
-                retry_deadline.map(|deadline| deadline.saturating_duration_since(Instant::now()));
-            if let Some(remaining) = attempt_timeout
-                && remaining.is_zero()
-            {
-                emit_initialize_metric(
-                    metric_transport,
-                    "error",
-                    attempt_count,
-                    /*retry_exhausted*/ false,
-                    "timeout",
-                );
-                let duration = timeout.unwrap_or(remaining);
-                return Err(anyhow!(
-                    "timed out handshaking with MCP server after {duration:?}"
-                ));
-            }
-
-            let transport = match pending_transport.take() {
-                Some(transport) => transport,
-                None => match Self::create_pending_transport(&self.transport_recipe).await {
-                    Ok(transport) => transport,
-                    Err(error) => {
-                        emit_initialize_metric(
-                            metric_transport,
-                            "error",
-                            attempt_count,
-                            /*retry_exhausted*/ false,
-                            "transport_create",
-                        );
-                        return Err(error);
-                    }
-                },
-            };
-
-            match Self::connect_pending_transport(
-                transport,
-                client_service.clone(),
-                attempt_timeout,
-            )
-            .await
-            {
-                Ok(result) => {
-                    emit_initialize_metric(
-                        metric_transport,
-                        "success",
-                        attempt_count,
-                        /*retry_exhausted*/ false,
-                        "none",
-                    );
-                    return Ok(result);
-                }
-                Err(error) if should_retry && Self::is_retryable_initialize_error(&error) => {
-                    let Some(retry_delay_ms) = retry_delay_ms else {
-                        emit_initialize_metric(
-                            metric_transport,
-                            "error",
-                            attempt_count,
-                            /*retry_exhausted*/ true,
-                            "retry_exhausted",
-                        );
-                        return Err(error);
-                    };
-                    let delay = Duration::from_millis(retry_delay_ms);
-                    warn!(
-                        attempt = attempt_count,
-                        max_attempts = STREAMABLE_HTTP_RETRY_DELAYS_MS.len() + 1,
-                        delay_ms = delay.as_millis(),
-                        error = %error,
-                        "streamable HTTP MCP initialize failed with a retryable error; retrying"
-                    );
-                    if !sleep_with_retry_deadline(delay, retry_deadline).await {
-                        emit_initialize_metric(
-                            metric_transport,
-                            "error",
-                            attempt_count,
-                            /*retry_exhausted*/ false,
-                            "timeout",
-                        );
-                        let duration = timeout.unwrap_or(delay);
-                        return Err(anyhow!(
-                            "timed out handshaking with MCP server after {duration:?}"
-                        ));
-                    }
-                }
-                Err(error) => {
-                    emit_initialize_metric(
-                        metric_transport,
-                        "error",
-                        attempt_count,
-                        /*retry_exhausted*/ false,
-                        "non_retryable",
-                    );
-                    return Err(error);
-                }
-            }
-        }
-
-        unreachable!("initialize retry loop should return on success or final error")
-    }
-
-    async fn run_service_operation<T, F, Fut>(
-        &self,
-        label: &str,
-        timeout: Option<Duration>,
-        operation: F,
-    ) -> Result<T>
-    where
-        F: Fn(Arc<RunningService<RoleClient, ElicitationClientService>>) -> Fut,
-        Fut: std::future::Future<Output = std::result::Result<T, rmcp::service::ServiceError>>,
-    {
-        let mut session_recovery_attempted = false;
-        let mut retry_attempt = 0;
-        let retry_deadline = timeout.map(|duration| Instant::now() + duration);
-
-        loop {
-            let service = self.service().await?;
-            let attempt_timeout =
-                retry_deadline.map(|deadline| deadline.saturating_duration_since(Instant::now()));
-            if let Some(remaining) = attempt_timeout
-                && remaining.is_zero()
-            {
-                let duration = timeout.unwrap_or(remaining);
-                return Err(ClientOperationError::Timeout {
-                    label: label.to_string(),
-                    duration,
-                }
-                .into());
-            }
-
-            match Self::run_service_operation_once(
-                Arc::clone(&service),
-                label,
-                attempt_timeout,
-                self.elicitation_pause_state.clone(),
-                &operation,
-            )
-            .await
-            {
-                Ok(result) => return Ok(result),
-                Err(error)
-                    if !session_recovery_attempted && Self::is_session_expired_404(&error) =>
-                {
-                    session_recovery_attempted = true;
-                    self.reinitialize_after_session_expiry(&service).await?;
-                }
-                Err(error)
-                    if Self::should_retry_tools_list_operation(label, retry_attempt, &error) =>
-                {
-                    let delay =
-                        Duration::from_millis(STREAMABLE_HTTP_RETRY_DELAYS_MS[retry_attempt]);
-                    retry_attempt += 1;
-                    warn!(
-                        label,
-                        attempt = retry_attempt,
-                        max_attempts = STREAMABLE_HTTP_RETRY_DELAYS_MS.len() + 1,
-                        delay_ms = delay.as_millis(),
-                        error = %error,
-                        "MCP service operation failed with a retryable error; retrying"
-                    );
-                    if !sleep_with_retry_deadline(delay, retry_deadline).await {
-                        let duration = timeout.unwrap_or(delay);
-                        return Err(ClientOperationError::Timeout {
-                            label: label.to_string(),
-                            duration,
-                        }
-                        .into());
-                    }
-                }
-                Err(error) => return Err(error.into()),
-            }
-        }
-    }
-
-    async fn run_service_operation_once<T, F, Fut>(
-        service: Arc<RunningService<RoleClient, ElicitationClientService>>,
-        label: &str,
-        timeout: Option<Duration>,
-        pause_state: ElicitationPauseState,
-        operation: &F,
-    ) -> std::result::Result<T, ClientOperationError>
-    where
-        F: Fn(Arc<RunningService<RoleClient, ElicitationClientService>>) -> Fut,
-        Fut: std::future::Future<Output = std::result::Result<T, rmcp::service::ServiceError>>,
-    {
-        match timeout {
-            Some(duration) => {
-                active_time_timeout(duration, pause_state.subscribe(), operation(service))
-                    .await
-                    .map_err(|_| ClientOperationError::Timeout {
-                        label: label.to_string(),
-                        duration,
-                    })?
-                    .map_err(ClientOperationError::from)
-            }
-            None => operation(service).await.map_err(ClientOperationError::from),
-        }
-    }
-
-    fn is_session_expired_404(error: &ClientOperationError) -> bool {
-        let ClientOperationError::Service(rmcp::service::ServiceError::TransportSend(error)) =
-            error
-        else {
-            return false;
-        };
-
-        error
-            .error
-            .downcast_ref::<StreamableHttpError<StreamableHttpClientAdapterError>>()
-            .is_some_and(|error| {
-                matches!(
-                    error,
-                    StreamableHttpError::Client(
-                        StreamableHttpClientAdapterError::SessionExpired404
-                    )
-                )
-            })
-    }
-
-    fn should_retry_tools_list_operation(
-        label: &str,
-        retry_attempt: usize,
-        error: &ClientOperationError,
-    ) -> bool {
-        label == "tools/list"
-            && retry_attempt < STREAMABLE_HTTP_RETRY_DELAYS_MS.len()
-            && Self::is_retryable_service_operation_error(error)
-    }
-
-    fn is_retryable_service_operation_error(error: &ClientOperationError) -> bool {
-        let ClientOperationError::Service(rmcp::service::ServiceError::TransportSend(error)) =
-            error
-        else {
-            return false;
-        };
-
-        error
-            .error
-            .downcast_ref::<StreamableHttpError<StreamableHttpClientAdapterError>>()
-            .is_some_and(Self::is_retryable_streamable_http_error)
-    }
-
-    fn is_retryable_initialize_error(error: &anyhow::Error) -> bool {
-        error.chain().any(|source| {
-            source
-                .downcast_ref::<HandshakeError>()
-                .is_some_and(|error| Self::is_retryable_client_initialize_error(&error.source))
-                || source
-                    .downcast_ref::<rmcp::service::ClientInitializeError>()
-                    .is_some_and(Self::is_retryable_client_initialize_error)
-        })
-    }
-
-    fn is_retryable_client_initialize_error(error: &rmcp::service::ClientInitializeError) -> bool {
-        match error {
-            rmcp::service::ClientInitializeError::TransportError { error, context }
-                if matches!(
-                    context.as_ref(),
-                    "send initialize request" | "send initialized notification"
-                ) =>
-            {
-                error
-                    .error
-                    .downcast_ref::<StreamableHttpError<StreamableHttpClientAdapterError>>()
-                    .is_some_and(Self::is_retryable_streamable_http_error)
-            }
-            _ => false,
-        }
-    }
-
-    fn is_retryable_streamable_http_error(
-        error: &StreamableHttpError<StreamableHttpClientAdapterError>,
-    ) -> bool {
-        match error {
-            StreamableHttpError::Client(
-                StreamableHttpClientAdapterError::RetryableHttpStatus(_)
-                | StreamableHttpClientAdapterError::HttpRequest(ExecServerError::HttpRequest(_)),
-            ) => true,
-            StreamableHttpError::Client(StreamableHttpClientAdapterError::HttpRequest(
-                ExecServerError::Server { code, message },
-            )) => {
-                *code == JSON_RPC_INTERNAL_ERROR_CODE && message.starts_with("http/request failed:")
-            }
-            _ => false,
-        }
     }
 
     async fn reinitialize_after_session_expiry(
